+++ discord/ui/checkbox.py
@@ -28,7 +28,7 @@ import os
 from typing import TYPE_CHECKING
 
 from ..components import Checkbox as CheckboxComponent
-from ..enums import ComponentType
+from ..enums import ComponentLimits, ComponentType
 from .item import ModalItem
 
 __all__ = ("Checkbox",)
@@ -105,8 +105,10 @@ class Checkbox(ModalItem):
     def custom_id(self, value: str):
         if not isinstance(value, str):
             raise TypeError(f"custom_id must be str not {value.__class__.__name__}")
-        if len(value) > 100:
-            raise ValueError("custom_id must be 100 characters or fewer")
+        if len(value) > ComponentLimits.custom_id_max.value:
+            raise ValueError(
+                f"custom_id must be {ComponentLimits.custom_id_max.value} characters or fewer"
+            )
         self.underlying.custom_id = value
 
     @property