+++ discord/ui/input_text.py
@@ -28,7 +28,7 @@ import os
 from typing import TYPE_CHECKING
 
 from ..components import InputText as InputTextComponent
-from ..enums import ComponentType, InputTextStyle
+from ..enums import ComponentLimits, ComponentType, InputTextStyle
 from .item import ModalItem
 
 __all__ = ("InputText", "TextInput")
@@ -101,16 +101,35 @@ class InputText(ModalItem):
         id: int | None = None,
     ):
         super().__init__()
-        if label and len(str(label)) > 45:
-            raise ValueError("label must be 45 characters or fewer")
-        if min_length and (min_length < 0 or min_length > 4000):
-            raise ValueError("min_length must be between 0 and 4000")
-        if max_length and (max_length < 0 or max_length > 4000):
-            raise ValueError("max_length must be between 1 and 4000")
-        if value and len(str(value)) > 4000:
-            raise ValueError("value must be 4000 characters or fewer")
-        if placeholder and len(str(placeholder)) > 100:
-            raise ValueError("placeholder must be 100 characters or fewer")
+        if label and len(str(label)) > ComponentLimits.text_input_label_max.value:
+            raise ValueError(
+                f"label must be {ComponentLimits.text_input_label_max.value} characters or fewer"
+            )
+        if min_length and (
+            min_length < ComponentLimits.text_input_min_length_min.value
+            or min_length > ComponentLimits.text_input_min_length_max.value
+        ):
+            raise ValueError(
+                f"min_length must be between {ComponentLimits.text_input_min_length_min.value} and {ComponentLimits.text_input_min_length_max.value}"
+            )
+        if max_length and (
+            max_length < ComponentLimits.text_input_max_length_min.value
+            or max_length > ComponentLimits.text_input_max_length_max.value
+        ):
+            raise ValueError(
+                f"max_length must be between {ComponentLimits.text_input_max_length_min.value} and {ComponentLimits.text_input_max_length_max.value}"
+            )
+        if value and len(str(value)) > ComponentLimits.text_input_max_length_max.value:
+            raise ValueError(
+                f"value must be {ComponentLimits.text_input_max_length_max.value} characters or fewer"
+            )
+        if (
+            placeholder
+            and len(str(placeholder)) > ComponentLimits.text_input_placeholder_max.value
+        ):
+            raise ValueError(
+                f"placeholder must be {ComponentLimits.text_input_placeholder_max.value} characters or fewer"
+            )
         if not isinstance(custom_id, str) and custom_id is not None:
             raise TypeError(
                 f"expected custom_id to be str, not {custom_id.__class__.__name__}"
@@ -180,8 +199,10 @@ class InputText(ModalItem):
     def custom_id(self, value: str):
         if not isinstance(value, str):
             raise TypeError(f"custom_id must be str not {value.__class__.__name__}")
-        if len(value) > 100:
-            raise ValueError("custom_id must be 100 characters or fewer")
+        if len(value) > ComponentLimits.custom_id_max.value:
+            raise ValueError(
+                f"custom_id must be {ComponentLimits.custom_id_max.value} characters or fewer"
+            )
         self.underlying.custom_id = value
 
     @property
@@ -193,8 +214,10 @@ class InputText(ModalItem):
     def label(self, value: str):
         if not isinstance(value, str):
             raise TypeError(f"label should be str not {value.__class__.__name__}")
-        if len(value) > 45:
-            raise ValueError("label must be 45 characters or fewer")
+        if len(value) > ComponentLimits.text_input_label_max.value:
+            raise ValueError(
+                f"label must be {ComponentLimits.text_input_label_max.value} characters or fewer"
+            )
         self.underlying.label = value
 
     @property
@@ -205,9 +228,13 @@ class InputText(ModalItem):
     @placeholder.setter
     def placeholder(self, value: str | None):
         if value and not isinstance(value, str):
-            raise TypeError(f"placeholder must be None or str not {value.__class__.__name__}")  # type: ignore
-        if value and len(value) > 100:
-            raise ValueError("placeholder must be 100 characters or fewer")
+            raise TypeError(
+                f"placeholder must be None or str not {value.__class__.__name__}"
+            )  # type: ignore
+        if value and len(value) > ComponentLimits.text_input_placeholder_max.value:
+            raise ValueError(
+                f"placeholder must be {ComponentLimits.text_input_placeholder_max.value} characters or fewer"
+            )
         self.underlying.placeholder = value
 
     @property
@@ -218,9 +245,16 @@ class InputText(ModalItem):
     @min_length.setter
     def min_length(self, value: int | None):
         if value and not isinstance(value, int):
-            raise TypeError(f"min_length must be None or int not {value.__class__.__name__}")  # type: ignore
-        if value and (value < 0 or value) > 4000:
-            raise ValueError("min_length must be between 0 and 4000")
+            raise TypeError(
+                f"min_length must be None or int not {value.__class__.__name__}"
+            )  # type: ignore
+        if value and (
+            value < ComponentLimits.text_input_min_length_min.value
+            or value > ComponentLimits.text_input_min_length_max.value
+        ):
+            raise ValueError(
+                f"min_length must be between {ComponentLimits.text_input_min_length_min.value} and {ComponentLimits.text_input_min_length_max.value}"
+            )
         self.underlying.min_length = value
 
     @property
@@ -231,9 +265,16 @@ class InputText(ModalItem):
     @max_length.setter
     def max_length(self, value: int | None):
         if value and not isinstance(value, int):
-            raise TypeError(f"min_length must be None or int not {value.__class__.__name__}")  # type: ignore
-        if value and (value <= 0 or value > 4000):
-            raise ValueError("max_length must be between 1 and 4000")
+            raise TypeError(
+                f"max_length must be None or int not {value.__class__.__name__}"
+            )  # type: ignore
+        if value and (
+            value < ComponentLimits.text_input_max_length_min.value
+            or value > ComponentLimits.text_input_max_length_max.value
+        ):
+            raise ValueError(
+                f"max_length must be between {ComponentLimits.text_input_max_length_min.value} and {ComponentLimits.text_input_max_length_max.value}"
+            )
         self.underlying.max_length = value
 
     @property
@@ -259,8 +300,10 @@ class InputText(ModalItem):
     def value(self, value: str | None):
         if value and not isinstance(value, str):
             raise TypeError(f"value must be None or str not {value.__class__.__name__}")  # type: ignore
-        if value and len(str(value)) > 4000:
-            raise ValueError("value must be 4000 characters or fewer")
+        if value and len(str(value)) > ComponentLimits.text_input_max_length_max.value:
+            raise ValueError(
+                f"value must be {ComponentLimits.text_input_max_length_max.value} characters or fewer"
+            )
         self.underlying.value = value
 
     @property