+++ CHANGELOG.md
@@ -14,6 +14,8 @@ These changes are available on the `master` branch, but have not yet been releas
 
 - Support for **Python 3.14**.
   ([#2948](https://github.com/Pycord-Development/pycord/pull/2948))
+- Added `ComponentLimits` and `EmbedLimits` enums for Discord API constraints.
+  ([#3217](https://github.com/Pycord-Development/pycord/pull/3217))
 
 ### Changed